+++ src/types.ts
@@ -1,6 +1,6 @@
 import { Box3, Camera, Sphere, Vector3, WebGLRenderer } from 'three';
 import { GetUrlFn, XhrRequest } from './loading/types';
-import { OctreeGeometry } from './loading2/OctreeGeometry';
+import { OctreeGeometry } from './loading2/octree-geometry';
 import { PointCloudOctree } from './point-cloud-octree';
 import { PointCloudOctreeGeometry } from './point-cloud-octree-geometry';
 import { LRU } from './utils/lru';
